+++ src/Symfony/Component/VarDumper/Dumper/CliDumper.php
@@ -33,12 +33,13 @@ class CliDumper extends AbstractDumper
         'default' => '0;38;5;208',
         'num' => '1;38;5;38',
         'const' => '1;38;5;208',
+        'virtual' => '3',
         'str' => '1;38;5;113',
         'note' => '38;5;38',
         'ref' => '38;5;247',
-        'public' => '',
-        'protected' => '',
-        'private' => '',
+        'public' => '39',
+        'protected' => '39',
+        'private' => '39',
         'meta' => '38;5;170',
         'key' => '38;5;113',
         'index' => '38;5;38',
@@ -347,7 +348,10 @@ class CliDumper extends AbstractDumper
             if ($cursor->hashKeyIsBinary) {
                 $key = $this->utf8Encode($key);
             }
-            $attr = ['binary' => $cursor->hashKeyIsBinary];
+            $attr = [
+                'binary' => $cursor->hashKeyIsBinary,
+                'virtual' => $cursor->attr['virtual'] ?? false,
+            ];
             $bin = $cursor->hashKeyIsBinary ? 'b' : '';
             $style = 'key';
             switch ($cursor->hashType) {
@@ -371,7 +375,7 @@ class CliDumper extends AbstractDumper
                     // no break
                 case Cursor::HASH_OBJECT:
                     if (!isset($key[0]) || "\0" !== $key[0]) {
-                        $this->line .= '+'.$bin.$this->style('public', $key).': ';
+                        $this->line .= '+'.$bin.$this->style('public', $key, $attr).': ';
                     } elseif (0 < strpos($key, "\0", 1)) {
                         $key = explode("\0", substr($key, 1), 2);
 
@@ -505,6 +509,9 @@ class CliDumper extends AbstractDumper
 
         if ('label' === $style && '' !== $value) {
             $value .= ' ';
+        }
+        if ($this->colors && ($attr['virtual'] ?? false)) {
+            $value = "\033[{$this->styles['virtual']}m".$value;
         }
 
         return $value;